requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
autobahn==24.4.2
Automat==24.8.1
boto3==1.34.114
botocore==1.34.114
certifi==2024.2.2
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.0.6
django-cors-headers==4.3.1
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
exponent_server_sdk==2.1.0
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
jmespath==1.0.1
multidict==6.0.5
numpy==2.1.1
openpyxl==3.1.5
pandas==2.2.2
pillow==10.3.0
polyline==2.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
service-identity==24.1.0
setuptools==74.0.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
zope.interface==7.0.3
